move funcparser.py and functions.py to lib/
