+++ pwndbg/lib/functions.py
